Added product delete confirmation page

+++ products/views.py
@@ -230,9 +230,18 @@ def delete_product(request, product_id):
         return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
-    product.delete()
-    messages.success(request, 'Product deleted!')
-    return redirect(reverse('products'))
+
+    if request.method == "POST":
+        product.delete()
+        messages.success(request, 'Product deleted!')
+        return redirect("home")
+    
+    template ='products/delete_product.html'
+    context = {
+        'product': product
+    }
+
+    return render(request, template, context)
 
 
 # Categories
